Fix Issue #9
Added an early return check that verifies that a `track` is the
`currently_playing_type` attribute in the `json_data` returned in
`getsong()`. This is a temporary fix to shut up the logger from
misthandling json for other types of media. This fixes issues with
podcast episodes playing as well.